LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/cross-origin-objects/cross-origin-objects-expected.txt

CONSOLE MESSAGE: line 1380: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "http://127.0.0.1:8800". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 1380: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "http://127.0.0.1:8800". Protocols, domains, and ports must match.

PASS Basic sanity-checking 
PASS Only whitelisted properties are accessible cross-origin 
PASS [[GetPrototypeOf]] should return null 
PASS [[SetPrototypeOf]] should throw 
PASS [[IsExtensible]] should return true for cross-origin objects 
PASS [[PreventExtensions]] should throw for cross-origin objects 
FAIL [[GetOwnProperty]] - Properties on cross-origin objects should be reported |own| Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "http://127.0.0.1:8800". Protocols, domains, and ports must match.
FAIL [[GetOwnProperty]] - Property descriptors for cross-origin properties should be set up correctly assert_equals: property descriptor for location should be configurable expected true but got false
PASS [[Delete]] Should throw on cross-origin objects 
PASS [[DefineOwnProperty]] Should throw for cross-origin objects 
FAIL [[Enumerate]] should return an empty iterator Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "http://127.0.0.1:8800". Protocols, domains, and ports must match.
PASS [[OwnPropertyKeys]] should return all properties from cross-origin objects 
PASS A and B jointly observe the same identity for cross-origin Window and Location 
PASS Cross-origin functions get local Function.prototype 
FAIL Cross-origin Window accessors get local Function.prototype undefined is not an object (evaluating 'f.name')
PASS Same-origin observers get different functions for cross-origin objects 
FAIL Same-origin obsevers get different accessors for cross-origin Window assert_true: different Window accessors per-incumbent script settings object expected true got false
FAIL Same-origin observers get different accessors for cross-origin Location Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "http://127.0.0.1:8800". Protocols, domains, and ports must match.
 
